requirements.txt
flask
python-dotenv
supabase-py
google-generativeai

